% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{gt_theme_quarto}
\alias{gt_theme_quarto}
\title{A gt Theme Generated by quartomonothemer}
\usage{
gt_theme_quarto(
  data,
  table.width = NULL,
  table.font.names = NULL,
  table.font.size = NULL,
  ...
)
}
\arguments{
\item{data}{A gt object}

\item{table.width}{Table Width}

\item{table.font.names}{Base font name}

\item{table.font.size}{Base font size}

\item{...}{`gt::tab_options` options}
}
\description{
A gt Theme Generated by quartomonothemer
}
