% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{theme_quarto}
\alias{theme_quarto}
\title{A ggplot2 Theme Generated by quartomonothemer}
\usage{
theme_quarto(font_title = NULL, font_text = NULL, size_base = NULL)
}
\arguments{
\item{font_title}{Title font}

\item{font_text}{Text font}

\item{size_base}{Base font size}
}
\value{
A ggplot2 theme
}
\description{
A ggplot2 Theme Generated by quartomonothemer
}
